update deps, fix ae2 and alloy forgery mixin closes #55 closes https://github.com/wisp-forest/alloy-forgery/issues/82 closes https://github.com/wisp-forest/alloy-forgery/issues/83

+++ module-compat/alloy-forgery/src/main/java/lol/bai/megane/module/alloyforgery/mixin/AccessorForgeControllerBlockEntity.java
@@ -12,7 +12,6 @@ public interface AccessorForgeControllerBlockEntity {
     ForgeDefinition getForgeDefinition();
 
     @Accessor
-    int getFuel();
-
+    float getFuel();
 
 }
